# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TAZGA Jewelry E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: collection
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, popular, price_asc, price_desc, bestselling] }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: listRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ─── Collections ─────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Orders ──────────────────────────────────────────────────────
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List active banners
      responses:
        "200":
          description: Banners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    patch:
      operationId: updateBanner
      tags: [banners]
      summary: Update banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete banner (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Newsletter ──────────────────────────────────────────────────
  /newsletter:
    get:
      operationId: listNewsletterSubscribers
      tags: [newsletter]
      summary: List newsletter subscribers (admin)
      responses:
        "200":
          description: Subscribers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsletterSubscriber"
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriber"
  /newsletter/{id}:
    delete:
      operationId: deleteNewsletterSubscriber
      tags: [newsletter]
      summary: Delete subscriber (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Cart ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, nameAr, price, images, categoryId, slug, inStock, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        shortDescription: { type: ["string", "null"] }
        shortDescriptionAr: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        price: { type: number }
        salePrice: { type: ["number", "null"] }
        costPrice: { type: ["number", "null"] }
        sku: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        categoryId: { type: ["integer", "null"] }
        collectionId: { type: ["integer", "null"] }
        category: { $ref: "#/components/schemas/Category", nullable: true }
        collection: { $ref: "#/components/schemas/Collection", nullable: true }
        tags: { type: array, items: { type: string } }
        material: { type: ["string", "null"] }
        weight: { type: ["string", "null"] }
        dimensions: { type: ["string", "null"] }
        inventory: { type: integer }
        inStock: { type: boolean }
        isFeatured: { type: boolean }
        isArchived: { type: boolean }
        averageRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        salesCount: { type: integer }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [name, nameAr, price, images]
      properties:
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        shortDescription: { type: string }
        shortDescriptionAr: { type: string }
        description: { type: string }
        descriptionAr: { type: string }
        price: { type: number }
        salePrice: { type: number }
        costPrice: { type: number }
        sku: { type: string }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        collectionId: { type: integer }
        tags: { type: array, items: { type: string } }
        material: { type: string }
        weight: { type: string }
        dimensions: { type: string }
        inventory: { type: integer }
        isFeatured: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        shortDescription: { type: string }
        shortDescriptionAr: { type: string }
        description: { type: string }
        descriptionAr: { type: string }
        price: { type: number }
        salePrice: { type: number }
        costPrice: { type: number }
        sku: { type: string }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        collectionId: { type: integer }
        tags: { type: array, items: { type: string } }
        material: { type: string }
        weight: { type: string }
        dimensions: { type: string }
        inventory: { type: integer }
        isFeatured: { type: boolean }
        isArchived: { type: boolean }
        inStock: { type: boolean }
    Collection:
      type: object
      required: [id, name, nameAr, slug, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"] }
        isFeatured: { type: boolean }
        isArchived: { type: boolean }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        productCount: { type: integer }
        createdAt: { type: string }
...
    Category:
      type: object
      required: [id, name, nameAr, slug, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"] }
        displayOrder: { type: integer }
        productCount: { type: integer }
        createdAt: { type: string }
    CategoryInput:
      type: object
      required: [name, nameAr, slug]
      properties:
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        description: { type: string }
        coverImage: { type: string }
        displayOrder: { type: integer }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        nameAr: { type: string }
        slug: { type: string }
        description: { type: string }
        coverImage: { type: string }
        displayOrder: { type: integer }
...
    Banner:
      type: object
      required: [id, title, isActive, displayOrder, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        titleAr: { type: ["string", "null"] }
        subtitle: { type: ["string", "null"] }
        subtitleAr: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        descriptionAr: { type: ["string", "null"] }
        buttonText: { type: ["string", "null"] }
        buttonUrl: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        displayOrder: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        createdAt: { type: string }
    BannerInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        titleAr: { type: string }
        subtitle: { type: string }
        subtitleAr: { type: string }
        description: { type: string }
        descriptionAr: { type: string }
        buttonText: { type: string }
        buttonUrl: { type: string }
        imageUrl: { type: string }
        isActive: { type: boolean }
        displayOrder: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
    BannerUpdate:
      type: object
      properties:
        title: { type: string }
        titleAr: { type: string }
        subtitle: { type: string }
        subtitleAr: { type: string }
        description: { type: string }
        descriptionAr: { type: string }
        buttonText: { type: string }
        buttonUrl: { type: string }
        imageUrl: { type: string }
        isActive: { type: boolean }
        displayOrder: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
    NewsletterSubscriber:
      type: object
      required: [id, email, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: ["string", "null"] }
        createdAt: { type: string }
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
        name: { type: string }
...