# lib/api-zod/src/generated/types/bannerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TAZGA Jewelry E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface BannerInput {
  title: string;
  titleAr?: string;
  subtitle?: string;
  subtitleAr?: string;
  description?: string;
  descriptionAr?: string;
  buttonText?: string;
  buttonUrl?: string;
  imageUrl?: string;
  isActive?: boolean;
  displayOrder?: number;
  startDate?: string;
  endDate?: string;
}