# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TAZGA Jewelry E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  coverImage?: string | null;
  isFeatured?: boolean;
  isArchived?: boolean;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  productCount?: number;
  createdAt: string;
}