# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TAZGA Jewelry E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { Collection } from './collection';

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  shortDescriptionAr?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /** @nullable */
  costPrice?: number | null;
  /** @nullable */
  sku?: string | null;
  images: string[];
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  collectionId?: number | null;
  category?: Category | null;
  collection?: Collection | null;
  tags?: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  dimensions?: string | null;
  inventory?: number;
  inStock: boolean;
  isFeatured?: boolean;
  isArchived?: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount?: number;
  salesCount?: number;
  createdAt: string;
}