# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TAZGA Jewelry E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  nameAr?: string;
  slug?: string;
  shortDescription?: string;
  shortDescriptionAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  salePrice?: number;
  costPrice?: number;
  sku?: string;
  images?: string[];
  categoryId?: number;
  collectionId?: number;
  tags?: string[];
  material?: string;
  weight?: string;
  dimensions?: string;
  inventory?: number;
  isFeatured?: boolean;
  isArchived?: boolean;
  inStock?: boolean;
}